requirements.txt
Django>=4.2
djangorestframework>=3.14
pytest>=7.0
pytest-django>=4.5
python-dotenv>=1.0
